Se cambia la estructura de las intefaces del proyecto
