+++ src/context/ui/UIContext.tsx
@@ -1,5 +1,5 @@
 "use client";
 import { createContext } from "react";
-import { ContextUIProps } from "../interfaces";
+import { ContextUIProps } from "../../utils/features/context/interfaces";
 
 export const UIContext = createContext({} as ContextUIProps);
