--- src/todos/interfaces/index.ts
@@ -1,12 +1,0 @@
-import { Todo as TodoPrisma } from '@prisma/client';
-
-// Components
-export interface Todos {
-  todos?: TodoPrisma[]
-}
-
-export interface ItemProps {
-  item?: TodoPrisma
-  toggleTodo: (id: string, done: boolean) => Promise<TodoPrisma>
-}
-
