+++ src/utils/features/shopping-cart/interfaces/index.ts
@@ -1,0 +1,15 @@
+import { ProductCardProps, ProductCookiesProps } from "@/utils/features/products/interfaces";
+
+export interface PropsCart {
+  product?: ProductCardProps;
+  quantity?: number;
+}
+
+export interface ShoppingCartProps {
+  totalItems: number;
+}
+
+export interface CheackoutProps {
+  items: ProductCookiesProps[];
+  labelButton: string;
+}
